Update modules resolution, move get_async_client to __init__.py
